Production test, Timer and Valve modules are modified
Production Test:
GetBarometerReading, GetPM620Reading, GetBarometerID functionality added

Valve:
Tostart one pulse TimeBaseStart need to add before enabling interrupt
Channe3 and Channel4 are not wroking.HAL Timer.c Modified to make it work

+++ Application/Sensor/DSensorTemperature.h
@@ -25,7 +25,7 @@ public:
     DSensorTemperature(void);
     eSensorError_t close();
     eSensorError_t initialise(void);
-    uint8_t GetTemperatureSensorDeviceID();
+    uint16_t GetTemperatureSensorDeviceID();
     eSensorError_t getTemperature(float *tempInCelcius);   
 };
 #endif /* __DSENSOR_BAROMETER_H */
